Add overrides to mixpanel
