+++ packages/back-end/src/types/Integration.ts
@@ -357,8 +357,8 @@ export interface SourceIntegrationInterface {
   getSensitiveParamKeys(): string[];
   getExperimentResultsQuery(
     snapshotSettings: ExperimentSnapshotSettings,
-    metrics: ExperimentMetricInterface[],
-    activationMetric: ExperimentMetricInterface | null,
+    metricDocs: ExperimentMetricInterface[],
+    activationMetricDoc: ExperimentMetricInterface | null,
     dimension: DimensionInterface | null
   ): string;
   getFormatDialect?(): FormatDialect;
